Improve the View Rendering
